step2. 기존 테스트코드 삭제
