+++ VendingMachineApp/VendingMachineApp/Models/Americano.swift
@@ -1,0 +1,8 @@
+//
+//  Americano.swift
+//  VendingMachineApp
+//
+//  Created by 김지선 on 2021/03/04.
+//
+
+import Foundation
