+++ VendingMachineApp/VendingMachineApp/Models/Beverage.swift
@@ -13,10 +13,10 @@ class Beverage: CustomStringConvertible {
     private let capacity: Int
     private let price: Int
     private let name: String
-    private let manufactureDate: Date
+    private let manufactureAt: Date
     
     var description: String {
-        return "\(brand), \(capacity)ml, \(price)원, \(name), \(manufactureDate.productDisplayFormat)"
+        return "\(brand), \(capacity)ml, \(price)원, \(name), \(manufactureAt.productDisplayFormat)"
     }
     
     init(brand: String, capacity: Int, price: Int, name: String, dateString: String){
@@ -24,22 +24,8 @@ class Beverage: CustomStringConvertible {
         self.capacity = capacity
         self.price = price
         self.name = name
-        self.manufactureDate = dateString.converToDate()
+        self.manufactureAt = Date.converToDate(from: dateString)
     }
 }
+ 
 
-extension String {
-    func converToDate() -> Date {
-        let manufactureDateFormatter = DateFormatter()
-        manufactureDateFormatter.dateFormat = "yyyyMMdd"
-        return manufactureDateFormatter.date(from: self)!
-    }
-}
-
-extension Date {
-    var productDisplayFormat: String {
-        let manufactureDateFormatter = DateFormatter()
-        manufactureDateFormatter.dateFormat = "yyyyMMdd"
-        return manufactureDateFormatter.string(from: self)
-    }
-}
