+++ VendingMachineApp/VendingMachineApp/Models/BeverageType.swift
@@ -1,0 +1,8 @@
+//
+//  BeverageType.swift
+//  VendingMachineApp
+//
+//  Created by 김지선 on 2021/03/05.
+//
+
+import Foundation
