+++ VendingMachineApp/VendingMachineAppTests/BeverageTests.swift
@@ -1,0 +1,8 @@
+//
+//  BeverageTests.swift
+//  VendingMachineAppTests
+//
+//  Created by 김지선 on 2021/03/08.
+//
+
+import Foundation
