+++ VendingMachineApp/VendingMachineAppTests/VendingMachineAppTests.swift
@@ -2,11 +2,10 @@
 //  VendingMachineAppTests.swift
 //  VendingMachineAppTests
 //
-//  Created by 김지선 on 2021/02/24.
+//  Created by 김지선 on 2021/03/08.
 //
 
 import XCTest
-@testable import VendingMachineApp
 
 class VendingMachineAppTests: XCTestCase {
 
@@ -25,7 +24,7 @@ class VendingMachineAppTests: XCTestCase {
 
     func testPerformanceExample() throws {
         // This is an example of a performance test case.
-        self.measure {
+        measure {
             // Put the code you want to measure the time of here.
         }
     }
